Fix resource loading
Most importantly: sizeof a data pointer is assumed to be 4, and fit into
int. Use intptr_t where an integer holding a pointer is expected. Use
pointer sizes in allocations. Somehow, nothing seems to break because of
gaps in structure arrays and such.

Enable Werr, but disable some warnings.

A DOOM1 wad seems to be loading fine, but Xlib stuff is still failing
(expecting to create a 8-bit depth PseudoColor screen).